Use policy in reply controller delete method

+++ app/Http/Controllers/RepliesController.php
@@ -26,6 +26,8 @@ class RepliesController extends Controller
 
     public function destroy(Reply $reply)
     {
+        $this->authorize('update', $reply);
+
         $reply->delete();
 
         return back()->with('flash', 'Reply deleted');
